fix(import): 按真实工资/办公水电模板修复导入 + 假数据(seed)可删
按用户真实 Excel(工资测试/办公室水电测试)实测定稿,修复三类问题:

工资(附表12):
- 列映射改真实叶子标签(基本工资/岗位工资/岗位技能津贴/高温及其他补贴/应出勤/请假…)
- role(职种/职务)文本列导入(matchByHeader 加 text 列,不再 cleanNum 变 0)
- 多月堆叠分段(标题 YYYY年M月工资表)→ ImportSummary 逐月路由(32+30 人,小计行排除)
- 请假小数 HALF_UP 取整(DTO 收 BigDecimal;0.5→1/2.8→3,已知限制)

办公水电(附表13/14):
- 列映射改真实表头(用电量(千瓦)/基准用电单价…);matchByHeader 前缀匹配扛单位后缀
- 月份/所属月份日期单元格:parseYearMonth 加 M/D/YY(年末位)解析(扛 Excel m/d/yy 格式)
- 跨年路由:行自带 acctMonth,2024+2025 同次导入各落各年;所属月份分离
- 导入=按(附表,月)upsert(uk_office 每月一行):覆盖被导入月(含种子),不波及未导入月

假数据可删:
- 去 seed 删除保护(Salary/OfficeService delete/batchDelete + 两屏表格 checkbox/删除按钮)
- 种子行现可勾选/单删/批删;clearImported 仍只删 import

matchByHeader: 标签前缀匹配(最长优先)+ text 列;source 类型补 'import'
IT 隔离: 删/改共享种子的用例加 @Transactional 回滚,跨年用例用无种子年份

验证: 全量 mvnw test 174 绿; 前端 build+vitest 111; 真后端+浏览器上传两个真实
xlsx 端到端(工资逐字段对照/办公水电 upsert 替换种子/seed 可删 UI),0 console error;
验后 DB remigrate 还原种子

+++ backend/src/main/java/com/park/demo3/dto/SalaryImportRequest.java
@@ -18,8 +18,8 @@ public record SalaryImportRequest(@NotNull List<Row> rows) {
         BigDecimal lunch,
         BigDecimal heat,
         BigDecimal commission,
-        Integer shouldDays,
-        Integer leaveDays,
+        BigDecimal shouldDays,   // 真实「应出勤（天）」;导入四舍五入取整(实体为 int)
+        BigDecimal leaveDays,    // 真实「请假（天）」含小数(2.125/0.5);导入四舍五入取整
         BigDecimal social,
         BigDecimal tax,
         BigDecimal otherDeduct
